attached_assets/requirements.txt
flet>=0.21.1
requests>=2.31.0
python-dotenv>=1.0.0 